refactor: move `ClientConfig` to integration test binary
