+++ bin/integration-tests/src/tests/mod.rs
@@ -1,4 +1,5 @@
 pub mod client;
+pub mod config;
 pub mod custom_transaction;
 pub mod fpi;
 pub mod network_transaction;
