+++ bin/integration-tests/src/tests/network_transaction.rs
@@ -15,17 +15,17 @@ use miden_client::assembly::{
 use miden_client::note::NoteTag;
 use miden_client::testing::common::{
     TestClient,
-    create_test_client,
     execute_tx_and_sync,
     insert_new_wallet,
     wait_for_blocks,
     wait_for_tx,
 };
-use miden_client::testing::config::ClientConfig;
 use miden_client::testing::note::NoteBuilder;
 use miden_client::transaction::{OutputNote, TransactionKernel, TransactionRequestBuilder};
 use miden_client::{Felt, ScriptBuilder, Word, ZERO};
 use rand::RngCore;
+
+use crate::tests::config::ClientConfig;
 
 // HELPERS
 // ================================================================================================
@@ -141,7 +141,7 @@ async fn get_counter_contract_account(
 
 pub async fn counter_contract_ntx(client_config: ClientConfig) -> Result<()> {
     const BUMP_NOTE_NUMBER: u64 = 5;
-    let (mut client, keystore) = create_test_client(client_config).await?;
+    let (mut client, keystore) = client_config.into_client().await?;
     client.sync_state().await?;
 
     let (network_account, library) =
@@ -204,7 +204,7 @@ pub async fn counter_contract_ntx(client_config: ClientConfig) -> Result<()> {
 }
 
 pub async fn recall_note_before_ntx_consumes_it(client_config: ClientConfig) -> Result<()> {
-    let (mut client, keystore) = create_test_client(client_config).await?;
+    let (mut client, keystore) = client_config.into_client().await?;
     client.sync_state().await?;
 
     let (network_account, library) =
