+++ crates/rust-client/src/test_utils/mod.rs
@@ -2,6 +2,3 @@ pub mod mock;
 
 #[cfg(feature = "std")]
 pub mod common;
-
-#[cfg(feature = "std")]
-pub mod config;
